Support complex file types
Some filetypes consist of combinations of several others. The way this
is usually supported is by sourcing the filetype scripts sequentially.
This leads to the filetype being set consecutively, which in turns calls
our logic several times in a row.

This commit supports having those complex filetypes override the
templates set by the component filetypes.
It begrudgingly pollutes the buffer namespace with a variable.
This was mainly tested on 'htmldjango'.